pipeline/requirements.txt
kafka-python==2.3.1
psycopg2-binary==2.9.9